refactor: Move server communication functions to module
